WIP on source chain files refactor
