+++ core/src/common/hash.rs
@@ -1,0 +1,2 @@
+#[derive(Clone, Debug, PartialEq)]
+pub struct Hash {}
