+++ core/src/common/mod.rs
@@ -1,1 +1,1 @@
-pub mod entry;
+pub mod hash;
